requirements.txt
Flask==2.0.1
google-generativeai==0.7.0
flask-cors==3.0.10
werkzeug==2.0.3